feat: 응원단 dialog 컴포넌트로 분리(#78)
